+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Fixed
 
+- Improve error handling for file writing (#31)
 - Wrong http status code in case of crash or missing executable (#27)
 
 ## [v0.4.2] - 2019-03-14